+++ RELEASE-NOTES.md
@@ -21,6 +21,7 @@ Project Page: https://github.com/a2stuff/a2d
 * Reduce cursor flickering while menus are drawn.
 * Reduce cursor flickering while list view rows are drawn.
 * Improve appearance of menu items with check marks.
+* Fix file picker behavior with more than 127 files.
 
 ### Launcher